add new prop to TopFridge to add search bar when necessary
