+++ rmr/components/TopFridge.tsx
@@ -1,10 +1,12 @@
 import React from "react";
+import SearchBar from "./SearchBar";
 
-export default function TopFridge({
-  children,
-}: {
+interface Props {
   children: React.ReactNode;
-}) {
+  showSearchBar?: boolean;
+}
+
+export default function TopFridge({ children, showSearchBar = true }: Props) {
   return (
     // add the top part of the fridge, the shine on the left and right, the handle, and the content in the middle
     <div
@@ -19,19 +21,26 @@ export default function TopFridge({
       <img
         src="/shine-left-crop.png"
         alt="shine"
-        className="absolute left-0 top-0 mt-3 ml-3 w-xs h-auto object-contain z-10"
+        className="absolute left-0 top-0 mt-1 ml-1 w-xs h-auto object-contain z-10"
       />
       <img
         src="/shine-right-crop.png"
         alt="shine"
-        className="absolute right-0 top-0 mt-3 mr-3 w-xs h-auto object-contain z-10"
+        className="absolute right-0 top-0 mt-1 mr-1 w-xs h-auto object-contain z-10"
       />
       {children}
 
       {/* Logo in left corner */}
-      <div className="absolute top-0 left-0 mt-[3.5em] ml-[6rem] w-[7rem] h-auto z-50">
+      <div className="absolute top-0 left-0 mt-[2.5rem] ml-[6rem] w-[7.5rem] h-auto z-50">
         <img src="/logo.png" alt="Logo" />
       </div>
+
+      {/* Only show the search bar if showSearchBar is true */}
+      {showSearchBar && (
+        <div className="absolute top-0 left-1/2 transform -translate-x-1/2 mt-[3.5rem] z-50">
+          <SearchBar />
+        </div>
+      )}
     </div>
   );
 }
